Fix Composition of WeatherView
