+++ lib/app.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_training/view/weather_view.dart';
+import 'package:flutter_training/view/weather_view/weather_view.dart';
 
 class App extends StatelessWidget {
   const App({super.key});
