--- lib/view/weather_view.dart
@@ -1,82 +1,0 @@
-import 'package:flutter/material.dart';
-
-class WeatherView extends StatelessWidget {
-  const WeatherView({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final textTheme = theme.textTheme;
-    final largeLabel = textTheme.labelLarge;
-
-    return Scaffold(
-      body: Center(
-        child: FractionallySizedBox(
-          widthFactor: 0.5,
-          child: Column(
-            children: [
-              const Spacer(),
-              Column(
-                children: [
-                  const AspectRatio(
-                    aspectRatio: 1,
-                    child: Placeholder(),
-                  ),
-                  const SizedBox(height: 16),
-                  Row(
-                    children: [
-                      Expanded(
-                        child: Text(
-                          '** ℃',
-                          textAlign: TextAlign.center,
-                          style: largeLabel?.copyWith(
-                            color: Colors.blue,
-                          ),
-                        ),
-                      ),
-                      Expanded(
-                        child: Text(
-                          '** ℃',
-                          textAlign: TextAlign.center,
-                          style: largeLabel?.copyWith(
-                            color: Colors.red,
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                  const SizedBox(height: 16),
-                ],
-              ),
-              Flexible(
-                child: Column(
-                  children: [
-                    const SizedBox(
-                      height: 80,
-                    ),
-                    Row(
-                      children: [
-                        Expanded(
-                          child: TextButton(
-                            onPressed: () {},
-                            child: const Text('Close'),
-                          ),
-                        ),
-                        Expanded(
-                          child: TextButton(
-                            onPressed: () {},
-                            child: const Text('Reload'),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ],
-                ),
-              )
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
